#359 Построение текстового описания результатов теперь находится в отдельном вспомогательным класса
