+++ Zilon.Core/Zilon.Core/Zilon.Core.csproj
@@ -169,6 +169,7 @@
     <Compile Include="Common\PropTags.cs" />
     <Compile Include="Schemes\SectorSubScheme.cs" />
     <Compile Include="Schemes\SectorSubSchemeMapFactory.cs" />
+    <Compile Include="Scoring\TextSummaryHelper.cs" />
     <Compile Include="Tactics\ActorInteractionBus.cs" />
     <Compile Include="Tactics\ActorInteractionEventBase.cs" />
     <Compile Include="Tactics\ActorInteractionEvents\DodgeActorInteractionEvent.cs" />
